refactor pytest test cases, test spangroup in a different file to avoid the issue when spacy doc handel entities and spangroups differently (likely the cause of the fails on some of the tests).
